Inital commit - Lottery.sol
The first major change to the way TBL operates - This contract no longer uses Oraclize (a third-party API) to generate the winning random number, this is due to a high failure rate in the past resulting in a locked contract. 
User input upon entry is now the key metric for random number generation.